+++ examples/nuxt-app/nuxt.config.ts
@@ -4,19 +4,15 @@ export default defineNuxtConfig({
   runtimeConfig: {
     public: {
       API_URL: '',
-      SITEID: 8888
+      tide: {
+        contentApi: {
+          baseUrl: 'https://develop.content.reference.sdp.vic.gov.au',
+          site: '8888'
+        }
+      }
     }
   },
   tide: {
-    contentApi: {
-      site: '8888',
-      baseUrl: process.env.CONTENT_API_SERVER,
-      apiPrefix: 'api/v1',
-      auth: {
-        username: 'dpc',
-        password: 'sdp'
-      }
-    },
     mapping: {
       content: {
         event: '@dpc-sdp/ripple-tide-event',
@@ -28,12 +24,6 @@ export default defineNuxtConfig({
       site: '@dpc-sdp/ripple-tide-api/mapping/site'
     },
     debug: false
-  },
-  vue: {
-    config: {
-      productionTip: false,
-      devtools: true
-    }
   },
   modules: [
     '@dpc-sdp/ripple-tide-api/nuxt',