+++ packages/ripple-tide-api/src/nuxt/components/Page.vue
@@ -1,15 +1,15 @@
 <template>
-  <slot v-if="page" :name="`${componentName}Page`" v-bind="{ page, site }">
+  <slot v-if="page && site" :name="`${componentName}Page`" v-bind="{ page, site }">
     <component :is="`${componentName}Page`" :page="page">
       <template #aboveHeader>
         <RplIconSprite />
         <slot name="aboveHeader">
-          <RplAlert v-for="alert in site.alerts" v-bind="alert" :key="alert.alertId" />
+          <RplAlert v-for="alert in site?.alerts" v-bind="alert" :key="alert.alertId" />
         </slot>
       </template>
       <template #primaryNav>
         <slot name="primaryNav">
-          <RplPrimaryNav v-bind="primaryNavProps"></RplPrimaryNav>
+          <RplPrimaryNav v-bind="primaryNavProps" :items="site?.menus.menuMain"></RplPrimaryNav>
         </slot>
       </template>
       <template #breadcrumbs>
@@ -18,15 +18,19 @@
         </slot>
       </template>
       <template #sidebar>
+        <slot name="aboveSidebar"></slot>
         <slot v-if="page.sidebar" name="sidebar">
-          <TideSidebarContactUs :contacts="page.sidebar.contacts" />
-          <TideSidebarRelatedLinks :items="page.sidebar.relatedLinks" />
-          <TideSidebarSocialShare :networks="page.sidebar.socialShareNetworks" :page-title="page.title" />
+          <TideSidebarSiteSectionNav v-if="page.sidebar.siteSectionNav" :nav="page.sidebar.siteSectionNav" />
+          <TideSidebarRelatedLinks v-if="page.sidebar.relatedLinks?.length" :items="page.sidebar.relatedLinks" />
+          <TideSidebarContactUs v-if="page.sidebar.contacts?.length" :contacts="page.sidebar.contacts" />
+          <TideSidebarSocialShare v-if="page.sidebar.socialShareNetworks?.length"
+            :networks="page.sidebar.socialShareNetworks" :page-title="page.title" />
         </slot>
+        <slot name="belowSidebar"></slot>
       </template>
       <template #footer>
         <slot name="footer">
-          <RplFooter :nav="site.menus.menuMain"></RplFooter>
+          <RplFooter :nav="site?.menus.menuMain"></RplFooter>
         </slot>
       </template>
     </component>
@@ -35,7 +39,7 @@
     <RplLayout>
       <template #aboveHeader>
         <RplIconSprite />
-        <slot v-if="site && site.alerts" name="aboveHeader">
+        <slot v-if="site && site?.alerts" name="aboveHeader">
           <RplAlert v-for="alert in site.alerts" v-bind="alert" :key="alert.alertId" />
         </slot>
       </template>
@@ -47,10 +51,11 @@
       <template #body>
         <!-- TODO: Add error handling in Error component -->
         <h1>{{ pageError.data?.error?.message }}</h1>
+        <p>There was an error</p>
       </template>
       <template #footer>
         <slot name="footer">
-          <RplFooter></RplFooter>
+          <RplFooter :items="site?.menus.menu"></RplFooter>
         </slot>
       </template>
     </RplLayout>
@@ -73,21 +78,23 @@ import { computed } from 'vue'
 import { pascalCase } from 'change-case'
 
 const route = useRoute()
-const config = useRuntimeConfig()
+const { public: config } = useRuntimeConfig()
+const siteId = config.tide?.contentApi.site
+
 // @ts-ignore
 const [{ data: site, error: siteError }, { data: page, error: pageError }] =
   await Promise.all([
     useFetch('/api/tide/site', {
       baseURL: config.API_URL || '',
       params: {
-        id: config.SITEID
+        id: siteId
       }
     }),
     useFetch('/api/tide/page', {
       baseURL: config.API_URL || '',
       params: {
         path: route.path,
-        site: config.SITEID
+        site: siteId
       }
     })
   ])