+++ packages/ripple-ui-core/src/index.ts
@@ -1,7 +1,6 @@
 // Lib
 export { default as rplEventBus } from './lib/eventbus'
 export { useAccessibleContainer } from './composables/useAccessibleContainer'
-export { default as registerRplCorePlugins } from './plugins/index'
 
 // Components
 export { default as RplAccordion } from './components/accordion/accordion.vue'
@@ -68,4 +67,4 @@ export { default as RplVerticalNavToggle } from './components/vertical-nav/toggl
 export { default as RplVerticalNavChildList } from './components/vertical-nav/child-list.vue'
 export { default as RplVerticalNavLink } from './components/vertical-nav/link.vue'
 
-export { default as getGrantStatus } from './utils/getGrantStatus.js'
+export { default as getGrantStatus } from './utils/getGrantStatus'