+++ packages/ripple-ui-core/src/plugins/index.ts
@@ -1,5 +1,0 @@
-import { registerRplSocialSharingPlugin } from './vue-social-sharing'
-
-export default (vueApp) => {
-  registerRplSocialSharingPlugin(vueApp)
-}