+++ packages/ripple-ui-core/src/plugins/vue-social-sharing.ts
@@ -1,6 +1,0 @@
-// TODO this should come with ripple-ui-core via nuxt plugin system
-import SocialSharing from 'vue-social-sharing'
-
-export const registerRplSocialSharingPlugin = (vueApp) => {
-  vueApp.use(SocialSharing)
-}